+++ R/add.R
@@ -9,7 +9,8 @@
 #' @examples "todo"
 add <- function(robopptx, content = NULL, slide = NA_integer_, shapeid = NA_integer_,
                 hint = NULL, ..., robo_class = NA_character_, r_class = NA_character_) {
-  stopifnot("robopptx" %in% class(my_layout))
+  # stopifnot("robopptx" %in% class(my_layout))
+  stop_if_not_robopptx(my_layout)
 
   dots <- list(...)
 
@@ -190,7 +191,9 @@ add_graph <- function(robopptx, graph = NULL, slide = NA_integer_,
 #' @export
 #' @examples "todo"
 add_slide <- function(robopptx, ...) {
-  stopifnot("robopptx" %in% class(robopptx))
+  # stopifnot("robopptx" %in% class(robopptx))
+  stop_if_not_robopptx(my_layout)
+
 
   args <- list(...)