+++ NEWS.md
@@ -1,5 +1,6 @@
 # robocop 0.1.0 (development version)
 
+* Add `file_open()` to open a file locally, for example, a PPTX file, using the default system application.
 * Add patched version of `officer::plot_layout_properties()`. There is a [PR pending](https://github.com/davidgohel/officer/pull/580). 
   This version, however, additionally plots the placeholder's shape id into its upper right corner.
 * Added `NEWS.md`, MIT license file